feat(ui): 新增自定义面板和插件管理对话框
- 添加了多个自定义面板类，包括数据解析面板、数据发送面板和波形图面板
- 实现了插件管理对话框，用于管理插件的启用、禁用和会话阻止
- 在 ui/__init__.py 中添加了新组件的导入和导出
