+++ main.py
@@ -1,20 +1,15 @@
 # main.py
+import re
 import struct
 import sys
 from datetime import datetime
 from pathlib import Path
 from typing import Optional, Dict, List, Any, Set  # Added Set
-import re
-
 from PySide6.QtCore import Slot, QByteArray, Qt, QEvent, QObject, Signal, QSettings
-from PySide6.QtGui import QAction, QTextCursor, QIcon, QIntValidator, QCloseEvent, QFont
+from PySide6.QtGui import QAction, QIcon, QCloseEvent
 from PySide6.QtWidgets import (
-    QApplication, QMainWindow, QWidget, QVBoxLayout, QHBoxLayout,
-    QGridLayout, QLabel, QComboBox, QLineEdit, QPushButton, QTextEdit,
-    QCheckBox, QMessageBox, QGroupBox, QScrollArea, QFileDialog,
-    QInputDialog, QDockWidget, QPlainTextEdit,
-    QDialog, QListWidget, QListWidgetItem,  # For Plugin Management Dialog
-    QSizePolicy
+    QApplication, QMainWindow, QWidget, QLabel, QLineEdit, QMessageBox, QFileDialog,
+    QInputDialog, QDockWidget, QDialog,  # For Plugin Management Dialog
 )
 
 # Core imports from your project structure
@@ -34,1095 +29,27 @@ from utils.data_models import SerialPortConfig, FrameConfig
 from utils.logger import ErrorLogger
 from utils.config_manager import ConfigManager
 from ui.theme_manager import ThemeManager
-from ui.widgets import ReceiveDataContainerWidget, SendDataContainerWidget
-
+from ui.widgets import SendDataContainerWidget
+from ui.fixed_panels import (
+    SerialConfigDefinitionPanelWidget,
+    CustomLogPanelWidget,
+    BasicCommPanelWidget,
+    ScriptingPanelWidget
+)
+from ui.adaptable_panels import (
+    AdaptedParsePanelWidget,
+    AdaptedSendPanelWidget,
+    AdaptedPlotWidgetPanel
+)
+from ui.dialogs import PluginManagementDialog
 from core.serial_manager import SerialManager
-from core.protocol_handler import ProtocolAnalyzer, FrameParser, get_data_type_byte_length, calculate_frame_crc16, \
+from core.protocol_handler import ProtocolAnalyzer, FrameParser, calculate_frame_crc16, \
     calculate_original_checksums_python
 from core.data_recorder import DataRecorder
-
 # Updated Plugin Architecture Imports
 from core.panel_interface import PanelInterface
 from core.plugin_manager import PluginManager  # Assumes plugin_manager_hot_reload_v2 is used
 
-
-# --- Plugin Management Dialog ---
-class PluginManagementDialog(QDialog):
-    """
-    A dialog for managing plugins: enabling/disabling and session-blocking.
-    """
-    plugin_status_changed_signal = Signal(str, str)  # module_name, new_status
-
-    def __init__(self, plugin_manager_ref: PluginManager, main_window_ref: 'SerialDebugger',
-                 parent: Optional[QWidget] = None):
-        super().__init__(parent)
-        self.plugin_manager = plugin_manager_ref
-        self.main_window_ref = main_window_ref
-        self.setWindowTitle("插件管理器")
-        self.setMinimumSize(600, 400)
-        self._init_ui()
-        self._populate_plugin_list()
-
-    def _init_ui(self):
-        layout = QVBoxLayout(self)
-
-        self.plugin_list_widget = QListWidget()
-        self.plugin_list_widget.setSizePolicy(QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Expanding)
-        layout.addWidget(self.plugin_list_widget)
-
-        buttons_layout = QHBoxLayout()
-        self.enable_button = QPushButton("启用选中插件")
-        self.enable_button.clicked.connect(lambda: self._change_plugin_status_action("enable"))
-        buttons_layout.addWidget(self.enable_button)
-
-        self.disable_button = QPushButton("禁用选中插件")
-        self.disable_button.clicked.connect(lambda: self._change_plugin_status_action("disable"))
-        buttons_layout.addWidget(self.disable_button)
-
-        self.session_block_button = QPushButton("会话阻止选中插件")
-        self.session_block_button.setToolTip("在当前应用会话中阻止此插件模块被加载，即使它被标记为已启用。")
-        self.session_block_button.clicked.connect(lambda: self._change_plugin_status_action("session_block"))
-        buttons_layout.addWidget(self.session_block_button)
-
-        self.unblock_button = QPushButton("取消会话阻止")
-        self.unblock_button.clicked.connect(lambda: self._change_plugin_status_action("unblock"))
-        buttons_layout.addWidget(self.unblock_button)
-
-        buttons_layout.addStretch()
-        self.refresh_button = QPushButton("刷新列表")
-        self.refresh_button.clicked.connect(self._populate_plugin_list)
-        buttons_layout.addWidget(self.refresh_button)
-
-        self.close_button = QPushButton("关闭")
-        self.close_button.clicked.connect(self.accept)
-        buttons_layout.addWidget(self.close_button)
-
-        layout.addLayout(buttons_layout)
-        self.setLayout(layout)
-
-    def _populate_plugin_list(self):
-        self.plugin_list_widget.clear()
-        discovered_plugins = self.plugin_manager.get_all_discovered_plugin_modules_metadata()
-
-        if not discovered_plugins:
-            self.plugin_list_widget.addItem("未发现任何插件模块。")
-            self.enable_button.setEnabled(False)
-            self.disable_button.setEnabled(False)
-            self.session_block_button.setEnabled(False)
-            self.unblock_button.setEnabled(False)
-            return
-
-        self.enable_button.setEnabled(True)
-        self.disable_button.setEnabled(True)
-        self.session_block_button.setEnabled(True)
-        self.unblock_button.setEnabled(True)
-
-        for plugin_meta in discovered_plugins:
-            module_name = plugin_meta.get("module_name", "未知模块")
-            display_name = plugin_meta.get("display_name", module_name.split('.')[-1])
-            version = plugin_meta.get("version", "N/A")
-            description = plugin_meta.get("description", "无描述。")
-            status = plugin_meta.get("status", "discovered")
-
-            item_text = f"{display_name} (v{version}) - {module_name}\n  状态: {status.upper()}\n  描述: {description}"
-            list_item = QListWidgetItem(item_text)
-            list_item.setData(Qt.ItemDataRole.UserRole, module_name)
-
-            font = list_item.font()
-            if status == "enabled":
-                list_item.setForeground(Qt.GlobalColor.darkGreen)
-            elif status == "disabled":
-                list_item.setForeground(Qt.GlobalColor.darkGray)
-                font.setItalic(True)
-            elif status == "blocklisted (session)":
-                list_item.setForeground(Qt.GlobalColor.red)
-                font.setStrikeOut(True)
-            list_item.setFont(font)
-
-            self.plugin_list_widget.addItem(list_item)
-
-    def _get_selected_module_name(self) -> Optional[str]:
-        current_item = self.plugin_list_widget.currentItem()
-        if current_item:
-            return current_item.data(Qt.ItemDataRole.UserRole)
-        return None
-
-    @Slot()
-    def _change_plugin_status_action(self, action_type: str):
-        module_name = self._get_selected_module_name()
-        if not module_name:
-            QMessageBox.warning(self, "操作插件", "请先从列表中选择一个插件模块。")
-            return
-
-        if action_type == "enable":
-            self.main_window_ref.update_plugin_enabled_status(module_name, True)
-            self.plugin_status_changed_signal.emit(module_name, "enabled")
-        elif action_type == "disable":
-            self.main_window_ref.update_plugin_enabled_status(module_name, False)
-            self.plugin_status_changed_signal.emit(module_name, "disabled")
-        elif action_type == "session_block":
-            self.main_window_ref.session_block_plugin_module(module_name)
-            self.plugin_status_changed_signal.emit(module_name, "session_blocked")
-        elif action_type == "unblock":
-            self.plugin_manager.unblock_module_for_session(module_name)
-            QMessageBox.information(self, "取消阻止",
-                                    f"模块 {module_name} 已从会话阻止列表中移除。\n如果之前已禁用，您可能需要重新启用它并通过“扫描/重载插件”来使其生效。")
-            self.plugin_status_changed_signal.emit(module_name, "unblocked_needs_scan")
-        self._populate_plugin_list()
-
-
-# --- Adapted Panel Widget Classes (Full Implementations) ---
-class AdaptedParsePanelWidget(PanelInterface):
-    PANEL_TYPE_NAME = "core_parse_panel"
-    PANEL_DISPLAY_NAME = "数据解析面板"
-
-    def __init__(self, panel_id: int, main_window_ref: 'SerialDebugger', initial_config: Optional[Dict] = None,
-                 parent: Optional[QWidget] = None):
-        super().__init__(panel_id, main_window_ref, initial_config, parent)
-        self.receive_data_containers: List[ReceiveDataContainerWidget] = []
-        self._init_ui()  # Initialize UI elements first
-        if initial_config:
-            self.apply_config(initial_config)
-        else:
-            if hasattr(self, 'parse_id_edit'):  # Ensure UI element exists
-                self.parse_id_edit.setText(f"C{self.panel_id}")
-            self._update_panel_title_from_parse_id()
-
-    def _init_ui(self):
-        layout = QVBoxLayout(self)
-        self.recv_display_group = QGroupBox()
-        recv_display_main_layout = QVBoxLayout()
-        recv_container_controls_layout = QHBoxLayout()
-        self.add_recv_container_button = QPushButton("添加显示项 (+)")
-        self.add_recv_container_button.clicked.connect(self._add_container_action_triggered)
-        recv_container_controls_layout.addWidget(self.add_recv_container_button)
-        self.remove_recv_container_button = QPushButton("删除显示项 (-)")
-        self.remove_recv_container_button.clicked.connect(self._remove_container_action_triggered)
-        self.remove_recv_container_button.setEnabled(False)
-        recv_container_controls_layout.addWidget(self.remove_recv_container_button)
-        recv_container_controls_layout.addStretch()
-        recv_display_main_layout.addLayout(recv_container_controls_layout)
-        self.scroll_area_recv_containers = QScrollArea()
-        self.scroll_area_recv_containers.setWidgetResizable(True)
-        self.recv_containers_widget = QWidget()
-        self.recv_containers_layout = QVBoxLayout(self.recv_containers_widget)
-        self.recv_containers_layout.setAlignment(Qt.AlignmentFlag.AlignTop)
-        self.scroll_area_recv_containers.setWidget(self.recv_containers_widget)
-        recv_display_main_layout.addWidget(self.scroll_area_recv_containers)
-        parse_config_layout = QGridLayout()
-        self.parse_id_label = QLabel("解析功能码(ID)[Hex]:")
-        parse_config_layout.addWidget(self.parse_id_label, 0, 0)
-        self.parse_id_edit = QLineEdit()
-        self.parse_id_edit.editingFinished.connect(self._update_panel_title_from_parse_id)
-        parse_config_layout.addWidget(self.parse_id_edit, 0, 1)
-        parse_config_layout.addWidget(QLabel("数据分配模式:"), 1, 0)
-        self.data_mapping_combo = QComboBox()
-        self.data_mapping_combo.addItems(["顺序填充 (Sequential)"])
-        parse_config_layout.addWidget(self.data_mapping_combo, 1, 1)
-        recv_display_main_layout.addLayout(parse_config_layout)
-        self.recv_display_group.setLayout(recv_display_main_layout)
-        layout.addWidget(self.recv_display_group)
-        self.setLayout(layout)
-        self._update_panel_title_from_parse_id()  # Call after all UI elements are created
-
-    def _update_panel_title_from_parse_id(self):
-        func_id_str = self.get_target_func_id().upper() if hasattr(self,
-                                                                   'parse_id_edit') and self.get_target_func_id() else "N/A"
-        new_title = f"{self.PANEL_DISPLAY_NAME} {self.panel_id} (ID: {func_id_str})"
-        if hasattr(self, 'recv_display_group'):
-            self.recv_display_group.setTitle(new_title)
-        self.dock_title_changed.emit(new_title)
-
-    @Slot()
-    def _add_container_action_triggered(self):
-        self.add_receive_data_container()
-
-    @Slot()
-    def _remove_container_action_triggered(self):
-        self.remove_receive_data_container()
-
-    def add_receive_data_container(self, config: Optional[Dict[str, Any]] = None, silent: bool = False) -> None:
-        container_id = self.main_window_ref.get_next_global_receive_container_id()
-        container = ReceiveDataContainerWidget(container_id, self.main_window_ref)
-        container.plot_target_changed_signal.connect(self.main_window_ref.handle_recv_container_plot_target_change)
-
-        if config:
-            container.name_edit.setText(config.get("name", f"RecvData_{container_id}"))
-            container.type_combo.setCurrentText(config.get("type", "uint8_t"))
-            if PYQTGRAPH_AVAILABLE and hasattr(container, 'plot_checkbox') and container.plot_checkbox:
-                container.plot_checkbox.setChecked(config.get("plot_enabled", False))
-        else:
-            container.name_edit.setText(f"RecvData_{container_id}")
-            if PYQTGRAPH_AVAILABLE and hasattr(container, 'plot_checkbox') and container.plot_checkbox:
-                container.plot_checkbox.setChecked(False)
-
-        self.recv_containers_layout.addWidget(container)
-        self.receive_data_containers.append(container)
-        self.remove_recv_container_button.setEnabled(True)
-
-        targets_for_dropdown = self.main_window_ref.get_available_plot_targets()
-        container.update_plot_targets(targets_for_dropdown)
-
-        if config and PYQTGRAPH_AVAILABLE and hasattr(container, 'plot_target_combo') and container.plot_target_combo:
-            plot_target_id = config.get("plot_target_id")
-            if plot_target_id is not None:
-                idx = container.plot_target_combo.findData(plot_target_id)
-                if idx != -1:
-                    container.plot_target_combo.setCurrentIndex(idx)
-            if hasattr(container, 'plot_checkbox'):
-                container.plot_target_combo.setEnabled(
-                    container.plot_checkbox.isChecked() and bool(targets_for_dropdown))
-        elif PYQTGRAPH_AVAILABLE and hasattr(container,
-                                             'plot_target_combo') and container.plot_target_combo and hasattr(container,
-                                                                                                              'plot_checkbox'):
-            container.plot_target_combo.setEnabled(container.plot_checkbox.isChecked() and bool(targets_for_dropdown))
-
-        if not silent and self.error_logger:
-            self.error_logger.log_info(f"Parse Panel {self.panel_id}: Added receive container {container_id}")
-
-    def remove_receive_data_container(self, silent: bool = False) -> None:
-        if self.receive_data_containers:
-            container_to_remove = self.receive_data_containers.pop()
-            container_id_removed = container_to_remove.container_id
-            self.recv_containers_layout.removeWidget(container_to_remove)
-            container_to_remove.deleteLater()
-            self.main_window_ref.clear_plot_curves_for_container(container_id_removed)
-            if not self.receive_data_containers: self.remove_recv_container_button.setEnabled(False)
-            if not silent and self.error_logger: self.error_logger.log_info(
-                f"Parse Panel {self.panel_id}: Removed receive container {container_id_removed}")
-
-    def get_target_func_id(self) -> str:
-        if hasattr(self, 'parse_id_edit'):
-            return self.parse_id_edit.text()
-        return ""
-
-    def dispatch_data(self, data_payload_ba: QByteArray) -> None:
-        if not self.receive_data_containers: return
-        current_offset = 0
-        parsed_data_for_log_export: Dict[str, str] = {}
-        timestamp_now = datetime.now()
-        if self.data_mapping_combo.currentText() == "顺序填充 (Sequential)":
-            for container_widget in self.receive_data_containers:
-                config = container_widget.get_config()
-                data_type = config["type"]
-                byte_len = get_data_type_byte_length(data_type)
-                segment = QByteArray()
-                if byte_len == -1:
-                    if current_offset < data_payload_ba.size(): segment = data_payload_ba.mid(
-                        current_offset); current_offset = data_payload_ba.size()
-                elif byte_len > 0:
-                    if current_offset + byte_len <= data_payload_ba.size(): segment = data_payload_ba.mid(
-                        current_offset, byte_len); current_offset += byte_len
-                container_widget.set_value(segment, data_type)
-                log_key_name = f"P{self.panel_id}_{config['name']}"
-                parsed_data_for_log_export[log_key_name] = container_widget.value_edit.text()
-                if PYQTGRAPH_AVAILABLE and config.get("plot_enabled", False) and config.get(
-                        "plot_target_id") is not None:
-                    target_plot_id = config["plot_target_id"]
-                    if target_plot_id in self.main_window_ref.dynamic_panel_instances:
-                        plot_panel_candidate = self.main_window_ref.dynamic_panel_instances[target_plot_id]
-                        if isinstance(plot_panel_candidate,
-                                      AdaptedPlotWidgetPanel) and plot_panel_candidate.plot_widget_container:
-                            val_float = container_widget.get_value_as_float()
-                            if val_float is not None: curve_name = f"P{self.panel_id}:{config['name']}"; plot_panel_candidate.update_data(
-                                config["id"], val_float, curve_name)
-        if parsed_data_for_log_export and hasattr(self.main_window_ref,
-                                                  'data_recorder'): self.main_window_ref.data_recorder.add_parsed_frame_data(
-            timestamp_now, parsed_data_for_log_export)
-
-    def get_config(self) -> Dict[str, Any]:
-        return {"parse_func_id": self.parse_id_edit.text() if hasattr(self, 'parse_id_edit') else "",
-                "data_mapping_mode": self.data_mapping_combo.currentText() if hasattr(self,
-                                                                                      'data_mapping_combo') else "顺序填充 (Sequential)",
-                "receive_containers": [c.get_config() for c in self.receive_data_containers]}
-
-    def apply_config(self, config: Dict[str, Any]):
-        if hasattr(self, 'parse_id_edit'): self.parse_id_edit.setText(config.get("parse_func_id", f"C{self.panel_id}"))
-        if hasattr(self, 'data_mapping_combo'): self.data_mapping_combo.setCurrentText(
-            config.get("data_mapping_mode", "顺序填充 (Sequential)"))
-        while self.receive_data_containers: self.remove_receive_data_container(silent=True)
-        for container_cfg in config.get("receive_containers", []): self.add_receive_data_container(config=container_cfg,
-                                                                                                   silent=True)
-        self._update_panel_title_from_parse_id()
-
-    def get_initial_dock_title(self) -> str:
-        func_id_text = self.parse_id_edit.text() if hasattr(self,
-                                                            'parse_id_edit') and self.parse_id_edit.text() else f"C{self.panel_id}"
-        return f"{self.PANEL_DISPLAY_NAME} {self.panel_id} (ID: {func_id_text.upper()})"
-
-    def update_children_plot_targets(self):
-        targets = self.main_window_ref.get_available_plot_targets()
-        for container in self.receive_data_containers:
-            current_target_id_data = None
-            if hasattr(container,
-                       'plot_target_combo') and container.plot_target_combo and container.plot_target_combo.count() > 0: current_target_id_data = container.plot_target_combo.currentData()
-            current_target_id = current_target_id_data if current_target_id_data is not None else None
-            container.update_plot_targets(targets)
-            if current_target_id is not None and hasattr(container,
-                                                         'plot_target_combo') and container.plot_target_combo:
-                idx = container.plot_target_combo.findData(current_target_id)
-                if idx != -1: container.plot_target_combo.setCurrentIndex(idx)
-            if hasattr(container, 'plot_checkbox') and hasattr(container,
-                                                               'plot_target_combo') and container.plot_checkbox and container.plot_target_combo: container.plot_target_combo.setEnabled(
-                container.plot_checkbox.isChecked() and bool(targets))
-
-    def on_panel_removed(self) -> None:
-        for container in self.receive_data_containers: self.main_window_ref.clear_plot_curves_for_container(
-            container.container_id)
-        if self.error_logger: self.error_logger.log_info(
-            f"Parse Panel {self.panel_id} removed. Cleaned up associated plot curves.")
-
-
-class AdaptedSendPanelWidget(PanelInterface):  # Full implementation
-    PANEL_TYPE_NAME = "core_send_panel"
-    PANEL_DISPLAY_NAME = "数据发送面板"
-
-    def __init__(self, panel_id: int, main_window_ref: 'SerialDebugger', initial_config: Optional[Dict] = None,
-                 parent: Optional[QWidget] = None):
-        super().__init__(panel_id, main_window_ref, initial_config, parent)
-        self.send_data_containers: List[SendDataContainerWidget] = []
-        self._next_local_send_container_id: int = 1
-        self._init_ui()
-        if initial_config:
-            self.apply_config(initial_config)
-        else:
-            if hasattr(self, 'panel_func_id_edit'): self.panel_func_id_edit.setText(f"S{self.panel_id:X}")
-            self._update_panel_title()
-
-    def _init_ui(self):
-        layout = QVBoxLayout(self)
-        self.send_data_group = QGroupBox()
-        send_data_main_layout = QVBoxLayout()
-        func_id_layout = QHBoxLayout()
-        func_id_layout.addWidget(QLabel("本面板功能码 (ID)[Hex,1B]:"))
-        self.panel_func_id_edit = QLineEdit()
-        self.panel_func_id_edit.setPlaceholderText("例如: C9")
-        self.panel_func_id_edit.editingFinished.connect(self._update_panel_title)
-        func_id_layout.addWidget(self.panel_func_id_edit)
-        send_data_main_layout.addLayout(func_id_layout)
-        send_container_controls_layout = QHBoxLayout()
-        self.add_button = QPushButton("添加发送项 (+)")
-        self.add_button.clicked.connect(self._add_container_action_triggered)
-        send_container_controls_layout.addWidget(self.add_button)
-        self.remove_button = QPushButton("删除发送项 (-)")
-        self.remove_button.clicked.connect(self._remove_container_action_triggered)
-        self.remove_button.setEnabled(False)
-        send_container_controls_layout.addWidget(self.remove_button)
-        send_container_controls_layout.addStretch()
-        send_data_main_layout.addLayout(send_container_controls_layout)
-        self.scroll_area = QScrollArea()
-        self.scroll_area.setWidgetResizable(True)
-        self.containers_widget = QWidget()
-        self.containers_layout = QVBoxLayout(self.containers_widget)
-        self.containers_layout.setAlignment(Qt.AlignmentFlag.AlignTop)
-        self.scroll_area.setWidget(self.containers_widget)
-        send_data_main_layout.addWidget(self.scroll_area)
-        self.send_frame_button_panel = QPushButton("发送此面板帧")
-        self.send_frame_button_panel.clicked.connect(self._trigger_send_frame)
-        self.send_frame_button_panel.setEnabled(self.main_window_ref.serial_manager.is_connected)
-        send_data_main_layout.addWidget(self.send_frame_button_panel)
-        self.send_data_group.setLayout(send_data_main_layout)
-        layout.addWidget(self.send_data_group)
-        self.setLayout(layout)
-        self._update_panel_title()
-
-    def _update_panel_title(self):
-        func_id_text = self.panel_func_id_edit.text().upper() if hasattr(self, 'panel_func_id_edit') else ""
-        title = f"{self.PANEL_DISPLAY_NAME} {self.panel_id} (ID: {func_id_text if func_id_text else '未定义'})"
-        if hasattr(self, 'send_data_group'): self.send_data_group.setTitle(title)
-        self.dock_title_changed.emit(title)
-
-    @Slot()
-    def _add_container_action_triggered(self):
-        self.add_send_data_container()
-
-    @Slot()
-    def _remove_container_action_triggered(self):
-        self.remove_send_data_container()
-
-    def add_send_data_container(self, config: Optional[Dict[str, Any]] = None, silent: bool = False):
-        container_id = self._next_local_send_container_id
-        container = SendDataContainerWidget(container_id, self.main_window_ref)
-        if config:
-            container.name_edit.setText(config.get("name", f"Data_{container_id}"))
-            container.type_combo.setCurrentText(config.get("type", "uint8_t"))
-            container.value_edit.setText(config.get("value", ""))
-        else:
-            container.name_edit.setText(f"Data_{container_id}")
-        self.containers_layout.addWidget(container)
-        self.send_data_containers.append(container)
-        self._next_local_send_container_id += 1
-        self.remove_button.setEnabled(True)
-        if not silent and self.error_logger: self.error_logger.log_info(
-            f"Send Panel {self.panel_id}: Added send data container {container_id}")
-
-    def remove_send_data_container(self, silent: bool = False):
-        if self.send_data_containers:
-            container_to_remove = self.send_data_containers.pop()
-            removed_id = container_to_remove.container_id
-            self.containers_layout.removeWidget(container_to_remove)
-            container_to_remove.deleteLater()
-            if not self.send_data_containers: self.remove_button.setEnabled(False)
-            if not silent and self.error_logger: self.error_logger.log_info(
-                f"Send Panel {self.panel_id}: Removed send data container {removed_id}")
-
-    @Slot()
-    def _trigger_send_frame(self):
-        if not self.main_window_ref.serial_manager.is_connected: QMessageBox.warning(self.main_window_ref, "警告",
-                                                                                     "串口未打开。"); return
-        panel_func_id_str = self.panel_func_id_edit.text()
-        if not panel_func_id_str: QMessageBox.warning(self.main_window_ref, "功能码缺失",
-                                                      "请输入此发送面板的目标功能码。"); return
-        self.main_window_ref.update_current_serial_frame_configs_from_ui()
-        final_frame = self.main_window_ref.assemble_custom_frame_from_send_panel_data(
-            panel_target_func_id_str=panel_func_id_str, panel_send_data_containers=self.send_data_containers)
-        if final_frame:
-            bytes_written = self.main_window_ref.serial_manager.write_data(final_frame)
-            if bytes_written == final_frame.size():
-                hex_frame_str = final_frame.toHex(' ').data().decode('ascii').upper()
-                msg = f"发送面板 {self.panel_id} (ID:{panel_func_id_str}) 发送 {bytes_written} 字节: {hex_frame_str}"
-                self.main_window_ref.status_bar_label.setText(msg)
-                if self.error_logger: self.error_logger.log_info(msg)
-                self.main_window_ref.protocol_analyzer.analyze_frame(final_frame, 'tx')
-                self.main_window_ref.data_recorder.record_raw_frame(datetime.now(), final_frame.data(),
-                                                                    f"TX (SendPanel {self.panel_id} ID:{panel_func_id_str})")
-                self.main_window_ref.append_to_custom_protocol_log_formatted(datetime.now(),
-                                                                             f"TX P{self.panel_id} ID:{panel_func_id_str}",
-                                                                             hex_frame_str, is_content_hex=True)
-            else:
-                if self.error_logger: self.error_logger.log_warning(
-                    f"Send Panel {self.panel_id}: Partial write. Expected {final_frame.size()}, wrote {bytes_written}")
-                self.main_window_ref.status_bar_label.setText(f"发送面板 {self.panel_id}: 写入错误")
-
-    def get_config(self) -> Dict[str, Any]:
-        return {"panel_func_id": self.panel_func_id_edit.text() if hasattr(self, 'panel_func_id_edit') else "",
-                "send_containers": [c.get_config() for c in self.send_data_containers]}
-
-    def apply_config(self, config: Dict[str, Any]):
-        if hasattr(self, 'panel_func_id_edit'): self.panel_func_id_edit.setText(
-            config.get("panel_func_id", f"S{self.panel_id:X}"))
-        while self.send_data_containers: self.remove_send_data_container(silent=True)
-        for c_cfg in config.get("send_containers", []): self.add_send_data_container(config=c_cfg, silent=True)
-        self._update_panel_title()
-
-    def get_initial_dock_title(self) -> str:
-        func_id_text = self.panel_func_id_edit.text().upper() if hasattr(self,
-                                                                         'panel_func_id_edit') and self.panel_func_id_edit.text() else f"S{self.panel_id:X}"
-        return f"{self.PANEL_DISPLAY_NAME} {self.panel_id} (ID: {func_id_text})"
-
-    def update_send_button_state(self, serial_connected: bool):
-        if hasattr(self, 'send_frame_button_panel'): self.send_frame_button_panel.setEnabled(serial_connected)
-
-
-if PYQTGRAPH_AVAILABLE and pg is not None:  # Full implementation
-    class AdaptedPlotWidgetPanel(PanelInterface):
-        PANEL_TYPE_NAME = "core_plot_widget_panel"
-        PANEL_DISPLAY_NAME = "波形图面板"
-
-        def __init__(self, panel_id: int, main_window_ref: 'SerialDebugger', initial_config: Optional[Dict] = None,
-                     parent: Optional[QWidget] = None):
-            super().__init__(panel_id, main_window_ref, initial_config, parent)
-            if initial_config:
-                self.plot_name = initial_config.get("name", f"波形图 {self.panel_id}")
-                self.max_data_points = initial_config.get("max_data_points", 300)
-            else:
-                self.plot_name = f"波形图 {self.panel_id}"
-                self.max_data_points = 300
-            self.curves: Dict[int, pg.PlotDataItem] = {}
-            self.data: Dict[int, Dict[str, list]] = {}
-            self.plot_widget_container: Optional[pg.PlotWidget] = None
-            self._init_ui()
-
-        def _init_ui(self):
-            layout = QVBoxLayout(self)
-            self.plot_widget_container = pg.PlotWidget(title=self.plot_name)
-            self.plot_widget_container.showGrid(x=True, y=True)
-            self.plot_widget_container.addLegend()
-            layout.addWidget(self.plot_widget_container)
-            controls_layout = QHBoxLayout()
-            self.rename_button = QPushButton("重命名图表")
-            self.rename_button.clicked.connect(self._rename_plot_action)
-            controls_layout.addWidget(self.rename_button)
-            self.clear_button = QPushButton("清空图表")
-            self.clear_button.clicked.connect(self.clear_plot)
-            controls_layout.addWidget(self.clear_button)
-            controls_layout.addStretch()
-            layout.addLayout(controls_layout)
-            self.setLayout(layout)
-
-        @Slot()
-        def _rename_plot_action(self):
-            current_name = self.plot_name
-            new_name, ok = QInputDialog.getText(self, "重命名波形图", "新名称:", QLineEdit.EchoMode.Normal,
-                                                current_name)
-            if ok and new_name and new_name != current_name:
-                self.plot_name = new_name
-                if self.plot_widget_container: self.plot_widget_container.setTitle(self.plot_name)
-                self.dock_title_changed.emit(self.plot_name)
-                self.main_window_ref.notify_plot_target_renamed(self.panel_id, self.plot_name)
-
-        def update_data(self, container_id: int, value: float, curve_name: str):
-            if not self.plot_widget_container: return
-            if container_id not in self.data: self.data[container_id] = {'x': [], 'y': []}; pen_color = pg.intColor(
-                len(self.curves) % 9, hues=9, values=1, alpha=200); self.curves[
-                container_id] = self.plot_widget_container.plot(name=curve_name, pen=pen_color)
-            self.data[container_id]['y'].append(value)
-            x_val = len(self.data[container_id]['y'])
-            self.data[container_id]['x'].append(x_val)
-            if len(self.data[container_id]['y']) > self.max_data_points: self.data[container_id]['y'].pop(0);
-            self.data[container_id]['x'].pop(0)
-            if container_id in self.curves: self.curves[container_id].setData(self.data[container_id]['x'],
-                                                                              self.data[container_id]['y'])
-
-        def clear_plot(self):
-            if not self.plot_widget_container: return
-            for cid in list(self.curves.keys()): self.remove_curve_for_container(cid, silent=True)
-            self.plot_widget_container.clear()
-            self.plot_widget_container.addLegend()
-            if self.error_logger: self.error_logger.log_info(
-                f"Plot Panel {self.panel_id} ('{self.plot_name}') cleared.")
-
-        def remove_curve_for_container(self, container_id: int, silent: bool = False):
-            if not self.plot_widget_container: return
-            if container_id in self.curves:
-                curve_to_remove = self.curves.pop(container_id)
-                self.plot_widget_container.removeItem(curve_to_remove)
-                self.data.pop(container_id, None)
-                if not silent and self.error_logger: self.error_logger.log_debug(
-                    f"[Plot Panel {self.panel_id}] Removed curve for container {container_id}.")
-
-        def get_config(self) -> Dict[str, Any]:
-            return {"name": self.plot_name, "max_data_points": self.max_data_points}
-
-        def apply_config(self, config: Dict[str, Any]):
-            self.plot_name = config.get("name", f"波形图 {self.panel_id}")
-            self.max_data_points = config.get("max_data_points", 300)
-            if self.plot_widget_container:
-                self.plot_widget_container.setTitle(self.plot_name)
-            self.dock_title_changed.emit(self.plot_name)
-
-        def get_initial_dock_title(self) -> str:
-            return self.plot_name
-
-        def on_panel_removed(self) -> None:
-            self.main_window_ref.notify_plot_target_removed(self.panel_id)
-            if self.error_logger: self.error_logger.log_info(
-                f"Plot Panel {self.panel_id} ('{self.plot_name}') removed.")
-else:  # Fallback if PYQTGRAPH_AVAILABLE is False
-    class AdaptedPlotWidgetPanel(PanelInterface):
-        PANEL_TYPE_NAME = "core_plot_widget_panel"
-        PANEL_DISPLAY_NAME = "波形图面板 (不可用)"
-
-        def __init__(self, panel_id: int, main_window_ref: 'SerialDebugger', initial_config: Optional[Dict] = None,
-                     parent: Optional[QWidget] = None):
-            super().__init__(panel_id, main_window_ref, initial_config, parent)
-            if initial_config:
-                self.plot_name = initial_config.get("name", f"波形图 {self.panel_id}")
-            else:
-                self.plot_name = f"波形图 {self.panel_id}"
-            self._init_ui()
-
-        def _init_ui(self):
-            layout = QVBoxLayout(self)
-            label = QLabel("PyQtGraph 未安装，波形图功能不可用。")
-            label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-            layout.addWidget(label); self.setLayout(layout)
-
-        def get_config(self):
-            return {"name": self.plot_name}
-
-        def apply_config(self, config):
-            self.plot_name = config.get("name", f"波形图 {self.panel_id}")
-            self.dock_title_changed.emit(self.plot_name + " (不可用)")
-
-        def get_initial_dock_title(self):
-            return self.plot_name + " (不可用)"
-
-        def update_data(self, container_id: int, value: float, curve_name: str):
-            pass
-
-        def clear_plot(self):
-            pass
-
-        def remove_curve_for_container(self, container_id: int, silent: bool = False):
-            pass
-
-
-class SerialConfigDefinitionPanelWidget(QWidget):  # Full implementation
-    connect_button_toggled = Signal(bool)
-    refresh_ports_requested = Signal()
-    config_changed = Signal()
-
-    def __init__(self, parent_main_window: 'SerialDebugger', parent: QWidget = None):
-        super().__init__(parent)
-        self.main_window_ref = parent_main_window
-        self.port_combo = QComboBox()
-        self.baud_combo = QComboBox()
-        self.data_bits_combo = QComboBox()
-        self.parity_combo = QComboBox()
-        self.stop_bits_combo = QComboBox()
-        self.refresh_ports_button = QPushButton("刷新")
-        self.connect_button = QPushButton("打开串口")
-        self.head_edit = QLineEdit()
-        self.saddr_edit = QLineEdit()
-        self.daddr_edit = QLineEdit()
-        self.id_edit = QLineEdit()
-        self.sum_check_display = QLineEdit()
-        self.add_check_display = QLineEdit()
-        self.checksum_mode_combo = QComboBox()
-        self._init_ui()
-
-    def _init_ui(self):
-        main_panel_layout = QVBoxLayout(self)
-        config_group = QGroupBox("串口配置")
-        config_layout = QGridLayout()
-        config_layout.addWidget(QLabel("端口:"), 0, 0)
-        self.port_combo.currentTextChanged.connect(self._emit_config_changed_if_valid_port)
-        config_layout.addWidget(self.port_combo, 0, 1)
-        config_layout.addWidget(QLabel("波特率:"), 1, 0)
-        self.baud_combo.addItems(["9600", "19200", "38400", "57600", "115200", "230400", "460800", "921600"])
-        self.baud_combo.setEditable(True)
-        if self.baud_combo.lineEdit(): self.baud_combo.lineEdit().setValidator(QIntValidator(0, 4000000, self))
-        self.baud_combo.currentTextChanged.connect(self._emit_config_changed)
-        config_layout.addWidget(self.baud_combo, 1, 1)
-        config_layout.addWidget(QLabel("数据位:"), 2, 0)
-        self.data_bits_combo.addItems(["8", "7", "6", "5"])
-        self.data_bits_combo.currentTextChanged.connect(self._emit_config_changed)
-        config_layout.addWidget(self.data_bits_combo, 2, 1)
-        config_layout.addWidget(QLabel("校验位:"), 3, 0)
-        self.parity_combo.addItems(["None", "Even", "Odd", "Space", "Mark"])
-        self.parity_combo.currentTextChanged.connect(self._emit_config_changed)
-        config_layout.addWidget(self.parity_combo, 3, 1)
-        config_layout.addWidget(QLabel("停止位:"), 4, 0)
-        self.stop_bits_combo.addItems(["1", "1.5", "2"])
-        self.stop_bits_combo.currentTextChanged.connect(self._emit_config_changed)
-        config_layout.addWidget(self.stop_bits_combo, 4, 1)
-        self.refresh_ports_button.clicked.connect(self.refresh_ports_requested.emit)
-        config_layout.addWidget(self.refresh_ports_button, 5, 0)
-        self.connect_button.setCheckable(True)
-        # self.connect_button.toggled.connect(self.connect_button_toggled.emit())
-        self.connect_button.clicked.connect(self._handle_internal_connect_button_click)  # <--- 新的连接方式
-
-        config_layout.addWidget(self.connect_button, 5, 1)
-        config_layout.setColumnStretch(1, 1)
-        config_group.setLayout(config_layout)
-        main_panel_layout.addWidget(config_group)
-        frame_def_group = QGroupBox("全局帧结构定义 (发送用)")
-        frame_def_layout = QGridLayout()
-        frame_def_layout.addWidget(QLabel("帧头(H)[Hex,1B]:"), 0, 0)
-        self.head_edit.setPlaceholderText("AA")
-        self.head_edit.editingFinished.connect(self._emit_config_changed)
-        frame_def_layout.addWidget(self.head_edit, 0, 1)
-        frame_def_layout.addWidget(QLabel("源地址(S)[Hex,1B]:"), 1, 0)
-        self.saddr_edit.setPlaceholderText("01")
-        self.saddr_edit.editingFinished.connect(self._emit_config_changed)
-        frame_def_layout.addWidget(self.saddr_edit, 1, 1)
-        frame_def_layout.addWidget(QLabel("目标地址(D)[Hex,1B]:"), 2, 0)
-        self.daddr_edit.setPlaceholderText("FE")
-        self.daddr_edit.editingFinished.connect(self._emit_config_changed)
-        frame_def_layout.addWidget(self.daddr_edit, 2, 1)
-        frame_def_layout.addWidget(QLabel("默认发送功能码(ID):"), 3, 0)
-        self.id_edit.setPlaceholderText("面板可覆盖, e.g., C0")
-        self.id_edit.editingFinished.connect(self._emit_config_changed)
-        frame_def_layout.addWidget(self.id_edit, 3, 1)
-        frame_def_layout.addWidget(QLabel("最后帧校验1/CRC高:"), 4, 0)
-        self.sum_check_display.setPlaceholderText("自动")
-        self.sum_check_display.setReadOnly(True)
-        frame_def_layout.addWidget(self.sum_check_display, 4, 1)
-        frame_def_layout.addWidget(QLabel("最后帧校验2/CRC低:"), 5, 0)
-        self.add_check_display.setPlaceholderText("自动")
-        self.add_check_display.setReadOnly(True)
-        frame_def_layout.addWidget(self.add_check_display, 5, 1)
-        frame_def_layout.addWidget(QLabel("校验模式:"), 6, 0)
-        self.checksum_mode_combo.addItem("原始校验 (Sum/Add)", ChecksumMode.ORIGINAL_SUM_ADD)
-        self.checksum_mode_combo.addItem("CRC-16/CCITT-FALSE", ChecksumMode.CRC16_CCITT_FALSE)
-        self.checksum_mode_combo.currentIndexChanged.connect(self._emit_config_changed)
-        frame_def_layout.addWidget(self.checksum_mode_combo, 6, 1)
-        frame_def_layout.setColumnStretch(1, 1)
-        frame_def_group.setLayout(frame_def_layout)
-        main_panel_layout.addWidget(frame_def_group)
-        main_panel_layout.addStretch(1)
-        self.setLayout(main_panel_layout)
-
-    def _emit_config_changed(self):
-        self.config_changed.emit()
-
-    def _emit_config_changed_if_valid_port(self, port_text: str):
-        if port_text != "无可用端口": self.config_changed.emit()
-
-    def update_ui_from_main_configs(self, serial_cfg: SerialPortConfig, frame_cfg: FrameConfig,
-                                    active_checksum: ChecksumMode):
-        widgets_to_block = [self.port_combo, self.baud_combo, self.data_bits_combo, self.parity_combo,
-                            self.stop_bits_combo, self.head_edit, self.saddr_edit, self.daddr_edit, self.id_edit,
-                            self.checksum_mode_combo]
-        for widget in widgets_to_block: widget.blockSignals(True)
-        if serial_cfg.port_name:
-            idx = self.port_combo.findData(serial_cfg.port_name)
-            if idx != -1:
-                self.port_combo.setCurrentIndex(idx)
-            else:
-                idx_text = self.port_combo.findText(serial_cfg.port_name, Qt.MatchFlag.MatchFixedString)
-            if idx_text != -1:
-                self.port_combo.setCurrentIndex(idx_text)
-            elif self.port_combo.count() > 0 and self.port_combo.itemText(0) != "无可用端口":
-                self.port_combo.setCurrentIndex(0)
-        baud_rate_str = str(serial_cfg.baud_rate)
-        if self.baud_combo.findText(baud_rate_str) == -1: self.baud_combo.addItem(baud_rate_str)
-        self.baud_combo.setCurrentText(baud_rate_str)
-        self.data_bits_combo.setCurrentText(str(serial_cfg.data_bits))
-        self.parity_combo.setCurrentText(serial_cfg.parity)
-        self.stop_bits_combo.setCurrentText(str(serial_cfg.stop_bits))
-        self.head_edit.setText(frame_cfg.head)
-        self.saddr_edit.setText(frame_cfg.s_addr)
-        self.daddr_edit.setText(frame_cfg.d_addr)
-        self.id_edit.setText(frame_cfg.func_id)
-
-        idx_cs = self.checksum_mode_combo.findData(active_checksum)
-        if idx_cs != -1:
-            self.checksum_mode_combo.setCurrentIndex(idx_cs)
-        else:
-            idx_def = -1  # Initialize idx_def
-            found_default_idx = self.checksum_mode_combo.findData(Constants.DEFAULT_CHECKSUM_MODE)
-            if found_default_idx != -1:
-                idx_def = found_default_idx
-
-            if idx_def != -1:
-                self.checksum_mode_combo.setCurrentIndex(idx_def)
-            # If default is also not found, combo box remains as is.
-
-        for widget in widgets_to_block: widget.blockSignals(False)
-
-    def get_serial_config_from_ui(self) -> SerialPortConfig:
-        port_name = None
-        port_name_text = self.port_combo.currentText()
-        if port_name_text != "无可用端口":
-            port_name_data = self.port_combo.currentData()
-            port_name = port_name_data if port_name_data is not None else None
-            if not port_name: port_name_text_parts = port_name_text.split(" "); port_name = port_name_text_parts[
-                0] if port_name_text_parts else None
-        try:
-            baud_rate = int(self.baud_combo.currentText())
-        except ValueError:
-            baud_rate = Constants.DEFAULT_BAUD_RATE
-        if hasattr(self.main_window_ref,
-                   'error_logger') and self.main_window_ref.error_logger: self.main_window_ref.error_logger.log_warning(
-            f"无效的波特率输入: '{self.baud_combo.currentText()}', 使用默认值 {baud_rate}"); self.baud_combo.setCurrentText(
-            str(baud_rate))
-        data_bits = int(self.data_bits_combo.currentText())
-        parity = self.parity_combo.currentText()
-        stop_bits_str = self.stop_bits_combo.currentText()
-        try:
-            stop_bits = float(stop_bits_str) if stop_bits_str == "1.5" else int(stop_bits_str)
-        except ValueError:
-            stop_bits = 1
-        if hasattr(self.main_window_ref,
-                   'error_logger') and self.main_window_ref.error_logger: self.main_window_ref.error_logger.log_warning(
-            f"无效的停止位输入: '{stop_bits_str}'，使用默认值 1")
-        return SerialPortConfig(port_name, baud_rate, data_bits, parity, stop_bits)
-
-    def get_frame_config_from_ui(self) -> FrameConfig:
-        return FrameConfig(head=self.head_edit.text().strip(), s_addr=self.saddr_edit.text().strip(),
-                           d_addr=self.daddr_edit.text().strip(), func_id=self.id_edit.text().strip())
-
-    def get_checksum_mode_from_ui(self) -> ChecksumMode:
-        mode = self.checksum_mode_combo.currentData()
-        if isinstance(mode, ChecksumMode):
-            return mode
-        return Constants.DEFAULT_CHECKSUM_MODE
-
-    def update_port_combo_display(self, available_ports: list[dict], current_port_name: Optional[str]):
-        self.port_combo.blockSignals(True)
-        self.port_combo.clear()
-        if not available_ports:
-            self.port_combo.addItem("无可用端口")
-            self.port_combo.setEnabled(False)
-            self.connect_button.setEnabled(False)
-        else:
-            for port_info in available_ports:
-                display_text = f"{port_info['name']} ({port_info.get('description', 'N/A')})"
-                self.port_combo.addItem(display_text, port_info['name'])
-            self.port_combo.setEnabled(True)
-            self.connect_button.setEnabled(True)
-            idx = -1
-            if current_port_name:
-                idx = self.port_combo.findData(current_port_name)
-            if idx != -1:
-                self.port_combo.setCurrentIndex(idx)
-            elif self.port_combo.count() > 0:
-                self.port_combo.setCurrentIndex(0)
-        self.port_combo.blockSignals(False)
-
-    def set_connection_status_display(self, connected: bool):
-        self.port_combo.setEnabled(not connected)
-        self.baud_combo.setEnabled(not connected)
-        self.data_bits_combo.setEnabled(not connected)
-        self.parity_combo.setEnabled(not connected)
-        self.stop_bits_combo.setEnabled(not connected)
-        self.refresh_ports_button.setEnabled(not connected)
-        self.connect_button.setChecked(connected)
-        self.connect_button.setText("关闭串口" if connected else "打开串口")
-        if not connected and (self.port_combo.count() == 0 or self.port_combo.currentText() == "无可用端口"):
-            self.connect_button.setEnabled(False)
-        elif not connected:
-            self.connect_button.setEnabled(True)
-
-    @Slot(bool) # <--- 新增的内部槽函数
-    def _handle_internal_connect_button_click(self, checked: bool):
-        """Internal slot to explicitly emit the connect_button_toggled signal."""
-        self.connect_button_toggled.emit(checked) # 明确用 checked 参数发射信号
-
-    def update_checksum_display(self, sum_check: str, add_check: str):
-        self.sum_check_display.setText(sum_check); self.add_check_display.setText(add_check)
-
-
-class CustomLogPanelWidget(QWidget):  # Full implementation
-    def __init__(self, main_window_ref: 'SerialDebugger', parent: Optional[QWidget] = None):
-        super().__init__(parent)
-        self.main_window_ref = main_window_ref
-        self.hex_checkbox = QCheckBox("Hex显示")
-        self.timestamp_checkbox = QCheckBox("时间戳")
-        self._init_ui()
-
-    def _init_ui(self):
-        layout = QVBoxLayout(self)
-        log_group = QGroupBox("自定义协议原始帧记录")
-        log_layout = QVBoxLayout()
-        self.log_text_edit = QTextEdit()
-        self.log_text_edit.setReadOnly(True)
-        self.log_text_edit.setFontFamily("Courier New")
-        log_layout.addWidget(self.log_text_edit)
-        options_layout = QHBoxLayout()
-        options_layout.addWidget(self.hex_checkbox)
-        options_layout.addWidget(self.timestamp_checkbox)
-        clear_button = QPushButton("清空记录区")
-        clear_button.clicked.connect(self.log_text_edit.clear)
-        options_layout.addWidget(clear_button)
-        options_layout.addStretch()
-        log_layout.addLayout(options_layout)
-        log_group.setLayout(log_layout)
-        layout.addWidget(log_group)
-        self.setLayout(layout)
-
-    def append_log(self, text: str): self.log_text_edit.moveCursor(
-        QTextCursor.MoveOperation.End); self.log_text_edit.insertPlainText(text); self.log_text_edit.moveCursor(
-        QTextCursor.MoveOperation.End)
-
-    def get_config(self) -> Dict: return {"hex_display": self.hex_checkbox.isChecked(),
-                                          "timestamp_display": self.timestamp_checkbox.isChecked()}
-
-    def apply_config(self, config: Dict): self.hex_checkbox.setChecked(
-        config.get("hex_display", False)); self.timestamp_checkbox.setChecked(config.get("timestamp_display", False))
-
-    def hex_checkbox_is_checked(self) -> bool: return self.hex_checkbox.isChecked()
-
-    def timestamp_checkbox_is_checked(self) -> bool: return self.timestamp_checkbox.isChecked()
-
-
-class BasicCommPanelWidget(QWidget):  # Full implementation
-    send_basic_data_requested = Signal(str, bool)
-
-    def __init__(self, main_window_ref: 'SerialDebugger', parent: Optional[QWidget] = None):
-        super().__init__(parent)
-        self.main_window_ref = main_window_ref
-        self.recv_hex_checkbox = QCheckBox("Hex显示")
-        self.recv_timestamp_checkbox = QCheckBox("显示时间戳")
-        self.send_hex_checkbox = QCheckBox("Hex发送")
-        self.send_button = QPushButton("发送")
-        self.receive_text_edit: Optional[QTextEdit] = None
-        self.send_text_edit: Optional[QLineEdit] = None
-        self._init_ui()
-        self._connect_signals()
-
-    def _init_ui(self):
-        main_layout = QVBoxLayout(self)
-        self._create_receive_group(main_layout)
-        self._create_send_group(main_layout)
-        self.setLayout(main_layout)
-
-    def _create_receive_group(self, main_layout: QVBoxLayout):
-        recv_group = QGroupBox("基本接收 (原始串行数据)")
-        recv_layout = QVBoxLayout()
-        self.receive_text_edit = QTextEdit()
-        self.receive_text_edit.setReadOnly(True)
-        font = QFont("Courier New", 10)
-        font.setStyleHint(QFont.StyleHint.Monospace)
-        self.receive_text_edit.setFont(font)
-        recv_layout.addWidget(self.receive_text_edit)
-        recv_options_layout = QHBoxLayout()
-        recv_options_layout.addWidget(self.recv_hex_checkbox)
-        recv_options_layout.addWidget(self.recv_timestamp_checkbox)
-        recv_options_layout.addStretch()
-        clear_basic_recv_button = QPushButton("清空接收区")
-        clear_basic_recv_button.clicked.connect(self.receive_text_edit.clear)
-        recv_options_layout.addWidget(clear_basic_recv_button)
-        recv_layout.addLayout(recv_options_layout)
-        recv_group.setLayout(recv_layout)
-        main_layout.addWidget(recv_group)
-
-    def _create_send_group(self, main_layout: QVBoxLayout):
-        send_group = QGroupBox("基本发送 (原始串行数据)")
-        send_layout = QVBoxLayout()
-        self.send_text_edit = QLineEdit()
-        self.send_text_edit.setPlaceholderText("输入要发送的文本或Hex数据 (如: AB CD EF 或 Hello)")
-        send_layout.addWidget(self.send_text_edit)
-        send_options_layout = QHBoxLayout()
-        send_options_layout.addWidget(self.send_hex_checkbox)
-        send_options_layout.addStretch()
-        clear_send_button = QPushButton("清空")
-        clear_send_button.clicked.connect(self.send_text_edit.clear)
-        send_options_layout.addWidget(clear_send_button)
-        self.send_button.clicked.connect(self._on_send_clicked)
-        send_options_layout.addWidget(self.send_button)
-        send_layout.addLayout(send_options_layout)
-        send_group.setLayout(send_layout)
-        main_layout.addWidget(send_group)
-
-    def _connect_signals(self):
-        if self.send_text_edit:
-            self.send_text_edit.returnPressed.connect(self._on_send_clicked)
-        self.send_hex_checkbox.toggled.connect(self._on_hex_mode_toggled)
-
-    @Slot(bool)
-    def _on_hex_mode_toggled(self, checked: bool):
-        if self.send_text_edit:
-            if checked:
-                self.send_text_edit.setPlaceholderText("输入Hex数据 (如: AB CD EF 或 ABCDEF)")
-            else:
-                self.send_text_edit.setPlaceholderText("输入要发送的文本")
-
-    @Slot()
-    def _on_send_clicked(self):
-        if not hasattr(self.main_window_ref, 'serial_manager'):
-            QMessageBox.warning(self, "错误",  "串口管理器未初始化。")
-            return
-        if not self.main_window_ref.serial_manager.is_connected:
-            QMessageBox.warning(self, "警告", "串口未打开，请先打开串口连接。")
-            return
-        if not self.send_text_edit:
-            return
-        text_to_send = self.send_text_edit.text().strip()
-        if not text_to_send:
-            QMessageBox.information(self, "提示", "请输入要发送的数据。")
-            return
-        is_hex = self.send_hex_checkbox.isChecked()
-        if is_hex and not self._validate_hex_input(text_to_send):
-            QMessageBox.warning(self, "输入错误", "Hex数据格式不正确。\n请输入有效的十六进制数据，如: AB CD EF 或 ABCDEF")
-            return
-
-        self.send_basic_data_requested.emit(text_to_send, is_hex)
-
-
-    def _validate_hex_input(self, text: str) -> bool:
-        if not text:
-            return False  # Handles empty string case explicitly
-
-        # hex_text will now always be assigned if text is not empty
-        hex_text = re.sub(r'[\s\-:,]', '', text.upper())
-
-        if not re.match(r'^[0-9A-F]*$', hex_text):
-            return False  # Handles non-hex characters
-
-        # Handles empty string after stripping (e.g., if input was just spaces)
-        # and ensures even length for valid hex.
-        if not hex_text:  # Check if hex_text became empty after re.sub (e.g. input was only spaces)
-            return False
-
-        return len(hex_text) % 2 == 0  # No need for `and len(hex_text) > 0` due to the `if not hex_text` check above
-
-    def append_receive_text(self, text: str):
-        if self.receive_text_edit is None:
-            return
-        cursor = self.receive_text_edit.textCursor()
-        cursor.movePosition(QTextCursor.MoveOperation.End)
-        self.receive_text_edit.setTextCursor(cursor)
-        self.receive_text_edit.insertPlainText(text)
-        self.receive_text_edit.ensureCursorVisible()
-    def set_send_enabled(self, enabled: bool):
-        self.send_button.setEnabled(enabled)
-        if self.send_text_edit:
-            self.send_text_edit.setEnabled(enabled)
-        self.send_hex_checkbox.setEnabled(enabled)
-    def get_config(self) -> Dict:
-        return {"recv_hex_display": self.recv_hex_checkbox.isChecked(),
-                "recv_timestamp_display": self.recv_timestamp_checkbox.isChecked(),
-                "send_hex_checked": self.send_hex_checkbox.isChecked()}
-
-    def apply_config(self, config: Dict):
-        self.recv_hex_checkbox.setChecked(config.get("recv_hex_display", False))
-        self.recv_timestamp_checkbox.setChecked(config.get("recv_timestamp_display", False))
-        self.send_hex_checkbox.setChecked(config.get("send_hex_checked", False))
-        self._on_hex_mode_toggled(self.send_hex_checkbox.isChecked())
-    def recv_hex_checkbox_is_checked(self) -> bool:
-        return self.recv_hex_checkbox.isChecked()
-
-    def recv_timestamp_checkbox_is_checked(self) -> bool:
-        return self.recv_timestamp_checkbox.isChecked()
-
-    def clear_receive_area(self):
-        if self.receive_text_edit: self.receive_text_edit.clear()
-
-    def clear_send_area(self):
-        if self.send_text_edit: self.send_text_edit.clear()
-
-    def get_send_text(self) -> str:
-        return self.send_text_edit.text() if self.send_text_edit else ""
-
-    def set_send_text(self, text: str):
-        if self.send_text_edit: self.send_text_edit.setText(text)
-
-
-class ScriptingPanelWidget(QWidget):  # Full implementation
-    execute_script_requested = Signal(str)
-
-    def __init__(self, main_window_ref: 'SerialDebugger', parent: Optional[QWidget] = None):
-        super().__init__(parent)
-        self.main_window_ref = main_window_ref
-        self.script_input_edit = QPlainTextEdit()
-        self.script_output_edit = QPlainTextEdit()
-        self._init_ui()
-
-    def _init_ui(self):
-        panel_layout = QVBoxLayout(self)
-        script_group = QGroupBox("脚本引擎")
-        group_layout = QVBoxLayout()
-        group_layout.addWidget(QLabel("脚本输入:"))
-        self.script_input_edit.setPlaceholderText("在此输入Python脚本...")
-        font = QFont("Courier New", 10)
-        font.setStyleHint(QFont.StyleHint.Monospace)
-        self.script_input_edit.setFont(font)
-        group_layout.addWidget(self.script_input_edit, 1)
-        self.run_script_button = QPushButton("执行脚本 (Run Script)")
-        self.run_script_button.clicked.connect(self._on_run_script_clicked)
-        group_layout.addWidget(self.run_script_button)
-        group_layout.addWidget(QLabel("脚本输出/结果:"))
-        self.script_output_edit.setReadOnly(True)
-        self.script_output_edit.setFont(font)
-        group_layout.addWidget(self.script_output_edit, 1)
-        clear_output_button = QPushButton("清空输出")
-        clear_output_button.clicked.connect(self.script_output_edit.clear)
-        group_layout.addWidget(clear_output_button)
-        script_group.setLayout(group_layout)
-        panel_layout.addWidget(script_group)
-        self.setLayout(panel_layout)
-
-    @Slot()
-    def _on_run_script_clicked(self):
-        script_text = self.script_input_edit.toPlainText()
-        if script_text.strip():
-            self.execute_script_requested.emit(script_text)
-        elif self.main_window_ref.error_logger:
-            self.main_window_ref.error_logger.log_warning("尝试执行空脚本。")
-            self.display_script_result("错误：脚本内容为空。")
-
-
-    def display_script_result(self, result_text: str):
-        self.script_output_edit.setPlainText(result_text)
-
-    def get_config(self) -> Dict:
-        return {"current_script": self.script_input_edit.toPlainText()}
-
-    def apply_config(self, config: Dict):
-        self.script_input_edit.setPlainText(config.get("current_script", ""))
 
 
 class SerialDebugger(QMainWindow):
